# backend/go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for SWD392 Chatbot RAG system.",
        "title": "SWD392 Chatbot RAG API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/documents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of all public documents with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "List public documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subject ID (UUID)",
                        "name": "subjectId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by document type ID (UUID)",
                        "name": "documentTypeId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by language ID (UUID)",
                        "name": "languageId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source ID (UUID)",
                        "name": "documentSourceId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (date_desc, date_asc, title_asc, title_desc, views_asc, views_desc)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 6)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/application.MyDocumentsDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get document, file, and chunk count statistics for the logged-in lecturer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lecturer"
                ],
                "summary": "Lecturer dashboard statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/application.DashboardSummaryDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/lookups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get dropdown listings of academic terms, subjects, types, sources, languages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metadata"
                ],
                "summary": "Get metadata lookups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/documents/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of documents owned by the logged-in lecturer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "List owned documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by subject ID (UUID)",
                        "name": "subjectId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by term ID (UUID)",
                        "name": "termId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by document type ID (UUID)",
                        "name": "documentTypeId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by language ID (UUID)",
                        "name": "languageId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source ID (UUID)",
                        "name": "documentSourceId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/application.MyDocumentsDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a document file (PDF/DOC/DOCX/PPT/PPTX) and start indexing",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Upload document file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Document file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Subject ID (UUID)",
                        "name": "subject_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Document Type ID (UUID)",
                        "name": "document_type_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Academic Term ID (UUID)",
                        "name": "academic_term_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Language ID (UUID)",
                        "name": "language_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Document Source ID (UUID)",
                        "name": "document_source_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Visibility (public, school_wide, private)",
                        "name": "visibility",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/application.DocumentCreateResultDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/documents/{slug}/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete document from database and S3 (only owner lecturer can delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "application.DocumentCreateResultDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "application.MyDocumentsDto": {
            "type": "object",
            "properties": {
                "active_upload_jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/application.UploadJobSummaryDto"
                    }
                },
                "approved_documents": {
                    "type": "integer"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/application.DocumentListItemDto"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "pending_documents": {
                    "type": "integer"
                },
                "rejected_documents": {
                    "type": "integer"
                },
                "total_chunks": {
                    "type": "integer"
                },
                "total_documents": {
                    "type": "integer"
                },
                "total_files": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}